pyproject.toml
[project]
name = "projeto-feic-antaq"
version = "0.1.0"
description = "Pipeline de dados do projeto FIEC e ANTAQ."
authors = [
    {name = "Zenilson Silva"}
]
readme = "README.md"
requires-python = "3.11.5"
dependencies = [
    "apache-airflow (>=2.10.5,<3.0.0)",
    "apache-airflow-providers-amazon (>=9.4.0,<10.0.0)",
    "duckdb (>=1.2.1,<2.0.0)",
    "minio (>=7.2.15,<8.0.0)",
    "boto3 (>=1.37.21,<2.0.0)",
    "pymssql (>=2.3.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
